% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My hello function}
\usage{
hello(name)
}
\arguments{
\item{name}{The name of the person}
}
\value{
The out from \code{\link{print}}
}
\description{
My hello function
}
\examples{
hello("Arafath")
\dontrun{
hello("Nihar")
}
}
